Move DashboardWidget to the widget module
This avoids cyclic dependencies.

JIRA: RAIL-3383
