+++ libs/sdk-ui-dashboard/src/layout/index.ts
@@ -1,7 +1,5 @@
 // (C) 2021 GoodData Corporation
-export { DashboardWidgetPropsProvider, useDashboardWidgetProps } from "./DashboardWidgetPropsContext";
 export { DashboardWidgetRenderer, IDashboardWidgetRendererProps } from "./DashboardWidgetRenderer";
 export { DashboardLayout } from "./DashboardLayout";
-export { DashboardLayoutProps, DashboardWidgetProps, CustomDashboardWidgetComponent } from "./types";
+export { DashboardLayoutProps } from "./types";
 export { DefaultDashboardLayout } from "./DefaultDashboardLayout";
-export { DefaultDashboardWidget, DefaultDashboardWidgetInner } from "./DefaultDashboardWidget";
