+++ libs/sdk-ui-dashboard/src/presentationComponents/DashboardItems/utils.ts
