+++ libs/sdk-ui-dashboard/src/widget/index.ts
@@ -1,3 +1,4 @@
 // (C) 2021 GoodData Corporation
 export * from "./insight";
 export * from "./kpi";
+export * from "./widget";
