+++ libs/sdk-ui-dashboard/src/widget/types.ts
@@ -1,3 +1,4 @@
 // (C) 2021 GoodData Corporation
 export { IDashboardInsightProps, CustomDashboardInsightComponent } from "./insight/types";
 export { DashboardKpiProps, CustomDashboardKpiComponent } from "./kpi/types";
+export { DashboardWidgetProps, CustomDashboardWidgetComponent } from "./widget/types";
