+++ libs/sdk-ui-dashboard/src/widget/widget/DashboardWidget.tsx
@@ -1,6 +1,6 @@
 // (C) 2020 GoodData Corporation
 import React from "react";
-import { useDashboardComponentsContext } from "../dashboardContexts";
+import { useDashboardComponentsContext } from "../../dashboardContexts";
 
 /**
  * @internal
