+++ libs/sdk-ui-dashboard/src/widget/widget/index.ts
@@ -1,0 +1,5 @@
+// (C) 2021 GoodData Corporation
+export { DashboardWidgetPropsProvider, useDashboardWidgetProps } from "./DashboardWidgetPropsContext";
+export { DashboardWidgetProps, CustomDashboardWidgetComponent } from "./types";
+export { DefaultDashboardWidget, DefaultDashboardWidgetInner } from "./DefaultDashboardWidget";
+export { DashboardWidget } from "./DashboardWidget";
